\name{ciaaw.mass.2016}
\docType{data}
\alias{ciaaw.mass.2016}
\title{Atomic masses of isotopes (IUPAC/CIAAW 2016)}
\description{
  This data set gives the atomic masses and uncertainties of all polyisotopic nuclides as they are used by the IUPAC/CIAAW.
}
\usage{ciaaw.mass.2016}
\format{A data frame with 268 rows and 4 variables:
	\itemize{
	\item {isotope: Symbol of the isotope}
	\item {element: Name of the element (lowercase english)}
	\item {mass: Atomic mass of the isotope in daltons}
	\item {uncertainty: Uncertainty of the atomic mass of the isotope as used by the IUPAC-CIAAW}
}}
\source{AME2016 \url{https://doi.org/10.1088/1674-1137/41/3/030003}, \url{http://www.ciaaw.org/}}
\keyword{datasets}