\name{ciaaw.mass}
\docType{data}
\alias{ciaaw.mass}
\title{Atomic masses of isotopes (IUPAC/CIAAW 2012)}
\description{
  This data set gives the atomic masses and uncertainties of all polyisotopic nuclides as they are used by the IUPAC/CIAAW.
}
\usage{ciaaw.mass}
\format{A data frame with 268 rows and 4 variables:
	\itemize{
	\item {isotope: Symbol of the isotope}
	\item {element: Name of the element (lowercase english)}
	\item {mass: Atomic mass of the isotope in daltons}
	\item {uncertainty: Uncertainty of the atomic mass of the isotope as used by the IUPAC-CIAAW}
}}
\source{\url{http://www.ciaaw.org/}}
\keyword{datasets}