\name{iridium.data}
\docType{data}
\alias{iridium.data}
\title{Iridium isotope ratio data from various studies}
\description{
  This data set gives the iridium isotope ratios as reported by various studies. These data are used by the IUPAC/CIAAW to determine the standard atomic weight of iridium.
}
\usage{iridium.data}
\format{A data frame.}
\source{IUPAC/CIAAW 2016}
\keyword{datasets}