\name{platinum.data}
\docType{data}
\alias{platinum.data}
\title{Platinum isotope ratio data from various studies}
\description{
  This data set gives the platinum isotope ratios as reported by various studies. These data are used by the IUPAC/CIAAW to determine the standard atomic weight of platinum.
}
\usage{platinum.data}
\format{A data frame.}
\source{IUPAC/CIAAW 2016}
\keyword{datasets}